src/api/requirements.txt
aiosqlite==0.17.0
anyio==3.6.1
argon2-cffi==21.3.0
asgiref==3.5.2
boto3==1.24.29
click==8.1.3
debugpy==1.6.0
h11==0.13.0
idna==3.3
sniffio==1.2.0
starlette==0.20.3
typing_extensions==4.2.0
uvicorn==0.17.6